backend/civic-issue-classifier/requirements.txt
streamlit
torch
torchvision
ultralytics
pillow
speechrecognition
deep-translator
requests
